chore(deps): Update gems and upgrade Puma to v7.0.0
Executed `bundle update` to keep the application's dependencies
up-to-date with their latest stable versions for security and
performance improvements.

This includes a major version upgrade for the Puma application server
to v7.0.0. Other gems have also been updated.

Refer to the `Gemfile.lock` diff for a complete list of changes.